refactor: seperate comment buttons from post features
