+++ src/entities/comment/ui/CommentDeleteButton.tsx
@@ -1,15 +1,15 @@
 import { Trash2 } from "lucide-react"
 import { Button } from "../../../shared/ui"
-import useComments from "../model/useComments"
+import useCommentButton from "../model/useCommentButton"
 
 interface CommentDeleteButtonProps {
   commentId: number
   postId: number
 }
 
-export const CommentDeleteButton = ({commentId, postId}:CommentDeleteButtonProps) => {
-  const { deleteCommentFetch } = useComments()
-
+export const CommentDeleteButton = ({ commentId, postId }: CommentDeleteButtonProps) => {
+  const {deleteCommentFetch} = useCommentButton()
+  
   return (
     <Button variant="ghost" size="sm" onClick={() => deleteCommentFetch(commentId, postId)}>
       <Trash2 className="w-3 h-3" />
