+++ src/features/comment/model/useComments.ts
@@ -40,48 +40,6 @@ const useComments = () => {
     }
   }
 
-  /** 댓글 좋아요 */
-  const likeCommentFetch = async (id: number, postId: number) => {
-    try {
-      const comment = comments[postId].find((c) => c.id === id)
-      if (!comment) {
-        throw new Error()
-      }
-      const response = await fetch(`/api/comments/${id}`, {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ likes: comment.likes + 1 }),
-      })
-      const data = await response.json()
-      setComments({
-        ...comments,
-        [postId]: comments[postId].map((comment) =>
-          comment.id === data.id ? { ...data, likes: comment.likes + 1 } : comment,
-        ),
-      })
-    } catch (error) {
-      console.error("댓글 좋아요 오류:", error)
-    }
-  }
-
-  /** 댓글 삭제 */
-  const deleteCommentFetch = async (id: number, postId: number) => {
-    try {
-      const { comments } = useCommentStore.getState()
-
-      await fetch(`/api/comments/${id}`, {
-        method: "DELETE",
-      })
-
-      setComments({
-        ...comments,
-        [postId]: comments[postId].filter((comment) => comment.id !== id),
-      })
-    } catch (error) {
-      console.error("댓글 삭제 오류:", error)
-    }
-  }
-
   /** 댓글 업데이트 */
   const updateCommentFetch = async () => {
     try {
@@ -113,18 +71,12 @@ const useComments = () => {
     setShowAddCommentDialog(true)
   }
 
-  /** 댓글 편집 */
-  const editComment = (comment: Comment) => {
-    setSelectedComment(comment)
-    setShowEditCommentDialog(true)
-  }
-
   /** 새 댓글 업데이트 */
   const updateNewComment = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     setNewComment({ ...newComment, body: e.target.value })
   }
 
-    /** 선택된 댓글 업데이트 */
+  /** 선택된 댓글 업데이트 */
   const updateSelectedComment = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     const newSelectedComment = selectedComment as Comment // TODO: 이렇게 해도 괜찮나?
     setSelectedComment({ ...newSelectedComment, body: e.target.value })
@@ -134,10 +86,7 @@ const useComments = () => {
     getComments,
     addCommentFetch,
     updateCommentFetch,
-    deleteCommentFetch,
-    likeCommentFetch,
     addComment,
-    editComment,
     newComment,
     updateNewComment,
     selectedComment,
